refactored functions into qc-helpers & s3.js
